kotlin-racing-game : [refactor] 클래스 및 상수 추출
- Input, Output Util 클래스 추출
- Car의 이동조건 범위 상수 추가
